update test env

+++ src/articles/router.py
@@ -13,8 +13,8 @@ router = APIRouter(
 )
 
 @router.get("/{article_id}", response_model=Article_api)
-async def read_article(article_id: int):
-    db_manager = DatabaseManager()
+async def read_article(article_id: int,
+                        db_manager: DatabaseManager = Depends(DatabaseManager)):
     article = await db_manager.get_article_by_id(article_id)
     await db_manager.close()
     if article is None:
@@ -22,17 +22,30 @@ async def read_article(article_id: int):
     return article
 
 @router.post("/", response_model=Article_api)
-async def create_article(article: ArticleBase, current_user: User_api = Depends(get_current_user)):
-    new_article = Article(title=article.title, content=article.content,
-                        author_id=current_user['id'], published_date=article.published_date)
-    db_manager = DatabaseManager()
+async def create_article(
+    article: ArticleBase, 
+    current_user: User_api = Depends(get_current_user),
+    db_manager: DatabaseManager = Depends(DatabaseManager)
+    ):
+    new_article = Article(
+        title=article.title, 
+        content=article.content,
+        author_id=current_user['id'], 
+        published_date=article.published_date
+    )
     await db_manager.add_article(new_article)
     await db_manager.close()
     return new_article
 
 @router.put("/{article_id}", response_model=Article_api)
-async def update_article(article_id: int, article: ArticleBase, current_user: User_api = Depends(get_current_user)):
-    db_manager = DatabaseManager()
+async def update_article(article_id: int, article: ArticleBase,
+                          current_user: User_api = Depends(get_current_user),
+                          db_manager: DatabaseManager = Depends(DatabaseManager)
+                          ): 
+    user_id = current_user['id']
+    author_id = await db_manager.get_user_id_by_article_id(article_id)
+    if current_user['role'] != 'admin' and user_id != author_id:
+        raise HTTPException(status_code=403, detail="Not enough permissions")
     updated_article = await db_manager.update_article(article_id, article)
     await db_manager.close()
     if updated_article is None:
@@ -40,8 +53,14 @@ async def update_article(article_id: int, article: ArticleBase, current_user: Us
     return updated_article
 
 @router.delete("/{article_id}", response_model=Article_api)
-async def delete_article(article_id: int, current_user: User_api = Depends(get_current_user)):
-    db_manager = DatabaseManager()
+async def delete_article(article_id: int,
+                        current_user: User_api = Depends(get_current_user),
+                        db_manager: DatabaseManager = Depends(DatabaseManager)):
+    
+    user_id = current_user['id']
+    author_id = await db_manager.get_user_id_by_article_id(article_id)
+    if current_user['role'] != 'admin' and user_id != author_id:
+        raise HTTPException(status_code=403, detail="Not enough permissions")
     deleted_article = await db_manager.delete_article(article_id)
     await db_manager.close()
     if deleted_article is None:
@@ -52,9 +71,9 @@ async def delete_article(article_id: int, current_user: User_api = Depends(get_c
 async def read_articles(page: int = Query(1, ge=1),
                         per_page: int = Query(10, ge=1, le=200),
                         author_id: Optional[int] = None,
-                        date_form: Optional[date] = None):
+                        date_form: Optional[date] = None,
+                        db_manager: DatabaseManager = Depends(DatabaseManager)):
 
-    db_manager = DatabaseManager()
     articles = await db_manager.get_articles(page, per_page, author_id, date_form)
     await db_manager.close()
     return articles
